Plot Two Figures Side-by-Side Identifying Common Elements

Plot two decoration graphs side-by-side identifying common nodes and common edges. This function is called by the function plot_compar.

side_plot(grph, dir, nd.var, focus = "nodes",
          nd.color = c("orange", "red"),
          nd.size = c(0.5, 1),
          ed.color = c("orange", "red"),
          ed.width = c(1, 2),
          lbl.size = 0.5)

Arguments

grph
List of two or more 'igraph' graphs created with the list_compar function.

dir
Working directory which contains the imgs, nodes, edges dataframes and the decoration images.

nd.var
Field of nodes on which the comparison will be done.

focus
Focus on nodes or on edges, by default focus = "nodes".

nd.color, nd.size, ed.color, ed.width
Graphical parameters for the nodes and edges. The different nodes/edges will be displayed with the first values of the vectors (eg, "orange") while the common nodes/edges will be displayed with the second values of the vectors (eg, "red").

lbl.size
Size of the labels

Value

No return value, group images side-by-side

See also